+++ local/src/main/java/com/jeanloth/project/android/kotlin/local/entities/RecipeWrapperEntity.kt
@@ -10,7 +10,7 @@ class RecipeWrapperEntity(
     @Id
     var ingredientWrapperId : Long = 0,
 
-    var quantity : Double = 0.0,
+    var quantity : Float = 0f,
     var quantityTypeLabel : String = IngredientQuantityType.KG.label
 
 ) : Entity {
